Module usbh::descriptor

Types for (standard) descriptors

This module contains types to represent various USB descriptors.

The parse submodule contains functions for parsing raw descriptors into these structures.

All descriptors have a common framing: the first two bytes contain the descriptor length and type respectively. This framing is represented by the Descriptor type.

To turn raw descriptor data into a Descriptor use the parse::any_descriptor function.

Such a descriptor can then be interpreted further, by examining the Descriptor::descriptor_type:

If the type matches one of the 5 standard types (TYPE_DEVICE, TYPE_CONFIGURATION, TYPE_STRING, TYPE_INTERFACE, TYPE_ENDPOINT), then it’s data can further be parsed by the respective methods in the parse module.

Otherwise it’s up to the driver to interpret the descriptor.

Modules

parse

Structs

ConfigurationAttributes

ConfigurationDescriptor

The configuration descriptor describes information about a specific device configuration.

Descriptor

Outer framing of a descriptor

DeviceDescriptor

A device descriptor describes general information about a USB device. It includes information that applies globally to the device and all of the device’s configurations. A USB device has only one device descriptor.

EndpointAddress

Address of an endpoint

EndpointAttributes

Attributes of an endpoint

EndpointDescriptor

Each endpoint used for an interface has its own descriptor.

InterfaceDescriptor

The interface descriptor describes a specific interface within a configuration. A configuration provides one or more interfaces, each with zero or more endpoint descriptors describing a unique set of endpoints within the configuration. When a configuration supports more than one interface, the endpoint descriptors for a particular interface follow the interface descriptor in the data returned by the GetConfiguration() request. An interface descriptor is always returned as part of a configuration descriptor. Interface descriptors cannot be directly accessed with a GetDescriptor() or SetDescriptor() request.

Enums

SynchronizationType

Synchronization type for an Isochronous endpoint

UsageType

Usage type for an Isochronous endpoint

Constants

TYPE_CONFIGURATION

descriptor_type identifying a ConfigurationDescriptor

TYPE_DEVICE

descriptor_type identifying a DeviceDescriptor

TYPE_ENDPOINT

descriptor_type identifying an EndpointDescriptor

TYPE_INTERFACE

descriptor_type identifying an InterfaceDescriptor

TYPE_STRING

descriptor_type identifying a StringDescriptor (not yet implemented)