Struct usbh::driver::kbd::KbdDriver

pub struct KbdDriver<const MAX_DEVICES: usize = 8> { /* private fields */ }

Driver for boot keyboards

By default, up to 8 connected keyboards can be handled. Events are reported for each device separately.

To increase (or decrease) the number of devices that can be handled, adjust the MAX_DEVICES parameter.

Note: the number of devices that can be handled also depends on UsbHost which limits the number of pipes that can be created. Each connected keyboard requires two pipes: a control pipe and an interrupt pipe.

Implementations

impl<const MAX_DEVICES: usize> KbdDriver<MAX_DEVICES>

pub fn new() -> Self

pub fn take_event(&mut self) -> Option<KbdEvent>

Returns the last keyboard event that occurred (if any) and clears it.

This method should be called directly after calling usb_host.poll(...).

Otherwise events may be lost.

For the meaning of events, please refer to the KbdEvent documentation.

pub fn set_idle<B: HostBus>( &mut self, dev_addr: DeviceAddress, latency: u8, host: &mut UsbHost<B> ) -> Result<(), KbdError>

Set interval for idle reports

If an idle interval is set, the keyboard will send out the current input report (i.e. pressed keys) regularly, even when no change to the pressed keys occurs.

The interval is expressed as a duration value, which is interpreted as a multiple of 4 ms.

Setting the duration to 0 disables idle reports. If they are disabled, input reports are only received when a key is pressed or released.

The USB HID specification recommends a default interval of 500ms for keyboards (duration value: 125).

pub fn set_led<B: HostBus>( &mut self, dev_addr: DeviceAddress, led: KbdLed, on: bool, host: &mut UsbHost<B> ) -> Result<(), KbdError>

Set the given KbdLed to the specified state.

The driver keeps track of the current output report (i.e. LED state basically) for each of the connected devices. Initially it is 0 (i.e. all LEDs are off).

This method updates one of the bits in the output report (identified by KbdLed) and sents the updated report to the device.